Adjust for strftime() handling %Y inconsistantly
On Debian systems, datetime.strftime("%Y") will produce less than a
4-digit year if the value is < 1000. On OS X and Windows sytems, it
will pad on the left with zeros to create a 4-digit year when the
value is < 1000.

I changed the code to detect whether padding was applied and supply
the correct padding if it was missing.

Also added recipe for a Docker container that could be used for
Ubuntu testing and development.

Bump version to 1.1.2